SlavePool: add MapperError exception
This commit adds a new exception class- MapperError that can be used to
report errors that arise during machine matching. In a previous patch
where I fixed corner cases of machine matching I used SlaveMachineError
which wasn't ideal since that is used for reporting errors in
SlaveMachine xml files. So I replaced it with the MapperError.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>