SlavePool: adjust how virtual matching is called
This commit modifies how we fallback to matching virtual machines when
normal matching fails.

First of all there is now a logging message informing the user about
this happening.

Next we need to check the interfaces of all the machines that are
requested by the recipe for parameters. Parameters are used for matching
and restricting possible interfaces, however this doesn't make sense
when matching virtual devices since these have no parameters to set. The
only exception is the parameter 'hwaddr' which can be used to specify
what mac address should the created interface have.

And finally I added a filter of the pool machines so that only virtual
machines with the parameter 'libvirt_domain' defined are considered when
matching virtual machines - this was a bug that went unnoticed since we
all use virtual machines all the time.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>