lnst-ctl: change meaning of -c option
This commit changes the meaning of the -c option. Previously this option
was used to perform config cleanup on machines before configuring them.
Since slaves are now capable of storing this configuration and perform
cleanup independently it is no longer needed.

The -c option is now used to specify an additional config file that
should be loaded after config files in standard places have been loaded.
This is a feature we agreed would be usefull for running automatic
tests.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>