NetTestSlave: cleanup when hello called
This commit moves the call of slave cleanup from the server loop to the
rpc method hello() as it makes more sense to do the cleanup whenever a
new session is started by the controller.

The rest of the patch changes related code on the controller side, this
is also related to the previous commit that removed the --cleanup option
and replaced it with --config.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>