lnst-ctl: Define/override aliases from CLI
This commit adds two CLI arguments to lnst-ctl:

    -a, --define-alias
    -A, --override-alias

Using these you can define or override an alias in the recipe you're
just executing. The difference between these two is in their effect.
Defied aliases (-a) can be overriden by the aliases that are defined in
the recie. The override aliases will replace all definitions of the
alias with the name in question within the entire recipe file.

Example:

    lnst-ctl -A http_port=80 recipe.xml run

Issue #1.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>