NetTestController: allow config_only for virtualized machines
This commit enables config_only functionality for virtualized machines.
This didn't work previously because the dynamically created devices
would be removed. This was needed because there was no way to store this
information between lnst-ctl runs.

Now the information is pickled and stored in a file /tmp/.lnst_virt_conf
this file is then detected on the next NetTestController object creation
(when the next recipe is being executed), and the configuration is
reversed removing interfaces, bridges, iptables rules and calling
machine_cleanup on slaves.

There are also a few helper get/set methods added in this commit.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>