Config: skip empty options
When a config file contained an option with an empty value this would
override the value loaded from previous config files and set it to
nothing. This resulted in some very weird behaviour when resource caches
were not being synchronized.

This commit fixes the problem by skipping options with no values

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>