SlavePool: add information about pool machines
This commit changes how LNST presents information about loaded pool
machines.

* the query messages were moved to debug logs
* added info messages about which pool directory is being processed
* after a directory is loaded, a colorized summary of availability is
  printed

I also adjusted what happens to virt machines when virtualization is
disabled but the machiens are available. In that case the libvirt_domain
attribute is ignored meaning that the machine will only be used in the
matching algorithm for static interfaces.

If a machine will be used in the dynamic interfaces matching algorithm
it will have its libvirt_domain printed in the summary described above.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>