Machine: exceptions during deconfiguration fix
This patch adds a few try-except blocks for machine deconfiguration. The
important part is that the slave is always disconnected in the end, and
that dynamic devices get detached when something fails.

Exceptions during deconfiguration caused that the controller never
disconnected from the slave, and the slave would be locked listening on
this socket. This would cause problems when there are multiple recipes
being processed in a single lnst-ctl call.

It is also important to continue the cleanup of remaining machines when
cleanup on one machine fails.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>