NetTestCommand: change exception handling
This commit fixes a bug that appears when a test option is missing in
the recipe. When such a situation appears, the slave throws an exception
that wasn't handled anywhere and resulted in the crash of the
controller.

After this commit the commands can either return properly- the have a
result, end with a KeyboardInterrupt or be killed - a default result is
set, or a different exception is raised - the command result is set to
fail and the exception is saved in result data. This shows up nicely in
results on the controller.

The missing option exception now also says which option is missing.

Another problem that appears is when an unexpected exception is raised
on the slave and is not handled- in such a situation the controller will
crash, and nothing nice can be written to the results. This situation is
now handled and the results will contain an exception logged into debug
logs (on the slave and sent to controller), and the controller will
report an unexpected exception being raised on a certain slave machine.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>