Machine: default command timeout
We used to have a default timeout for commands with the previous rpc
implementation, but didn't implement it in the new one. This commit adds
a default timeout for commands, the time is set in the DEFAULT_TIMEOUT
constant, which for now is 60 seconds.

This also fixes result reporting for killed commands.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>