NetTestController: print value of slave exceptions
Previously the controller would only print that there was an exception
raised on a certain slave machine, and there would be a traceback in
debug logs from that slave. Now the value of the exception will be added
to our controller logs, that makes it easier to spot what went wrong
without looking up the traceback.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>