test_modules: Removing the 'Test' prefix from modules
In order to be loaded properly, each test module was required to use
'Test' prefix in its name. However, in recipes, the prefix was never
used.

This was justified when the modules were mixed with other code. However,
this hasn't been the case for quite a while now. The modules are kept in
their own directory, so requiring a prefix is not necessary any more.

This patch removes the prefixes from module names.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>