recipe format: change <machines> to <network>
This patch set changes the tag <machines> to <network>, we also changed
the <machine> tag to <host> and the 'network' attribute used in <eth>
tags has been changed to 'label'.

The 'network' attribute was changed for slave machine xml files as well.

The commit also contains appropriate changes to the recipes. I tested
them for parsability, functinality will be checked later.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>