Multicast: fix result data format
The Multicast test module was using names that couldn't be transformed
into valid xml element names which resulted in the test failing. This
commit fixes that by replacing spaces with underscores.

I also fixed the regular expression- the left part of it sometimes
contained '-' or '.' which resulted in some lines not being correctly
parsed.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>