team recipes: bug fixes
This patch fixes a few bugs that prevented some team recipes from
working.

The labels in "interfaces-team_lb.xml" are obvious- other files used
different ones.

I increased the wait time for "sequence_iperf.xml" because it sometimes
took longer for the iperf server to bind to the address.

I moved the information about the "testifc" in the "sequence_pktgen.xml"
to be the last one, because for some reason the pktgen won't work
properly if that interface is not the last one added for some recipes.

There were a few other problems that I've encountered but they were all
related to weird race conditions that were probably caused by the fact
that I'm using virtual machines. I'm leaving those unchanged, because
there isn't really any valid solution for them.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>