recipes: Changing the 'machine' attr to 'host'
The machines tag were recently changed to <network> and with that, we
also renamed machines to hosts. This patch changes the attributes of
commands in tasks to match these changes.

The 'machine' attribute was changed to 'host' for the <config>,
<run>, and the signal tags (<wait>, <intr>, <kill>).

The commit also contains the same changes for the python task API.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>