schema-recipe: task can't be empty
The scheme allowed the existence of empty tasks, which would later cause
an expcetion during the recipe execution.

This patch fixes the schema so that there is always at least one command
in the schema. The solution isn't perfect since now the error reported
is something like "expected element <run>" which isn't entirely true
(for e.g. <config> can be found as well), but it will do for now.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>