move Xml* modules from Common/ to Controller/
These modules are only used by the controller so they should be placed
in the Controller/ directory. We originally placed them in Common/
because we expected external tools to use them and have them dependent
on the lnst-common package, however now it seems that having them
dependent on the lnst-ctl package will be better. The reason behind it
is that if you're working with recipes in an external tool you're
probably also running lnst-ctl on the same machine.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>