Logs: use machine names instead of IP addresses
This commit changes the format of log messages, instead of using
hostnames, which are usually IP addresses, we will use the host ids that
are specified in the supplied recipe.

This makes it more obvious what's happening on which machine. The
information about the used hosts is still available at the beginning of
recipe execution when the controller finds a suitable match.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>