NmConfigDevice: add detection of NM version
This commit adds a new function that returns the version of the running
NM daemon. The version is determined through dbus. For now I've used it
to detect whether to manually remove soft network devices or to leave NM
do it (NM version >= 0.9.9 removes them automatically). Later this will
be used to determine if team devices are supported by NM.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>