NmConfigDevice: add team configuration through NM
This commit adds support for configuring team devices through NM. The
support is limited to NM versions >= 0.9.9.
The code is similar to how bond configuration is done -> we create
connection objects for the team device and all the slaves, remove them
in reverse order. (De)Activation is done in the same order.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>