VirtualInterface: add support for setting device driver
Using the libvirt API binding allows us to create dynamic devices with
specific device drivers easily. This commit adds that functionality.

To specify a device driver you need to set a parameter of your
dynamically created device in your recipe like this:
<eth ...>
    ...
    <params>
        <param name="driver" value="e1000"/>
    </params>
    ...
</eth>

Keep in mind that this will be used as a matching restriction for
matching of physical devices, and when matching virtual machines the
value needs to be supported by libvirt. An invalid value will result in
an exception being raised.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>