ConnectionHandler: retrieve ALL messages from connections
The check_connections method only recieved one message per file
descriptor. This wasn't a problem because the other messages would be
retrieved in the next iteration. However this way we will be able to
order the recieved messages, so that messages from the controller can be
handled last, after all recieved updates (e.g. netlink).

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>