NmConfigDevice: InterfaceManager integration
The NmConfigDevice module now supports the new InterfaceManager module
instead of the old NetConfig way. I also refactored certain variable
names so that they make more sense and are consistent with the rest of
the code.

In addition to this the NetworkManager code needed reimplementing in
some places to work with the new program flow of device configuration
and to support more use cases. The most important part of this is that
configuration of master interfaces (e.g. bonding or team) doesn't also
configure the slave interfaces. Instead it changes finds their
ConfigDevice object and changes it's configuration so that the device
is enslaved on activation.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>