Machine, NetTestController: ordered interface configuration
This commit completes the master/slave hierarchy of machine interfaces
and adds a new method get_ordered_interfaces to the Machine class that
returns the machines interfaces in order of slaves before masters. The
interfaces are then configured in this order and after all devices are
configured they are set UP in the same order. The reverse ordering is
used in the Machine class when deconfiguring devices.

Interface information sent to the slave will also contain the id of the
master interface or None. This will be used by the NmConfig classes to
activate master connections.

This is required for both NM and iproute2 to work properly.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>