NetTestController: cleanup of saved configuration waits for results
When cleaning up a saved configuration the controller will wait for a
result message from the slave when calling the "machine_cleanup" method.
Without this the controller would sometimes remove the virtual devices
before the slave had the time to deconfigure them which will cause an
exception on the slave.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>