NetTestSlave: fix cleanup
This commit removes the state variable _ctl_clean_exit. The original
idea was to use this variable to differentiate between the config_only
controller run from a controller crash/disconnection. However this posed
more problems than it solved so now the slave will simply do a complete
machine cleanup every time the "hello" method is called or when the
slave is terminated.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>