NetTestController: always save config_only configuration
Until now we only saved config_only configuration if the run was
virtualized, this was so that we could nicely deconfigure dynamically
created network devices. This commit makes it so that the configuration
is stored for every config_only run, even if virtualization wasn't used.
This is not required for functionality since slaves will automatically
deconfigure themselves when needed, but it can be used to call
deconfiguration without running a different config_only run or run an
entire recipe.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>