NetTestController: sync only required resources
From now the controller will only synchronize resources required by the
slave machine for recipe execution. This is done by traversing the
commands that will be executed and searching for those that require any
kind of resource.

This commit also removes a piece of code from the function
_prepare_command that handled raising exceptions when a tool was not
found, this is now handled in the code that traverses the recipe
commands.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>