NetTestSlave: resource_table init
When the resource_table is initialized or cleared it should have at
least empty 'module' and 'tools' subdictionaries created. This is so
that proper exceptions are raised in case nothing was synchronized to
the slave - before this commit an exception KeyError 'module not
defined' would be raised, instead an exception "module/tool xyz not
found" is more appropriate.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>