lnst-ctl: add -r, --reduce-sync option
This option disables the full resource synchronization for python tasks.
Instead it relies on the user to use the sync_resources HostAPI method
to manually synchronize required modules and tools.

The option doesn't have any effect on recipes that use xml tasks, these
will always be automatically synchronized in reduced mode when only
required resources are copied to slaves.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>