XmlParser: save full filename in element objects
I'm not exactly sure why we would only store the full path in the root
element and the following elements would only use the basename of the
file. This comit changes that so that the absolute path is stored in all
element objects.

This will be usefull when figuring out the point of reference for
non xinclude attributes that refer to files.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>