NetTestController: add support for {module,tools}_dir
This commit adds support for searching additional resource directories
specified by the module_dir and tools_dir attributes.

We create a deep copy of the original resource table and then update it
with the newly found modules or tools. After the task is processed we
restore the original resource table.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>