fix incorrect handling of config command failure
The patch fixes an unhandled exception in case a config command fails.
Instead of overwriting res_data we have to append to it since other keys
are expected to be there, even empty. Further, NetTestSlave expects that
run() method returns a result that's referenced later. I've also changed
"msg" key to "err_msg" since the latter is used in NetTestSlave.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>