ExecCmd: exec_cmd processes should close fds
When creating a subprocess using the subprocess.Popen method, the newly
created process inherits the file descriptors of the parent by default.
This includes the server socket on the Slave machines which can cause
problems if the slave starts a daemon process (for example ovs-vswitchd)
that is not stopped when lnst-slave dies.

This commit fixes that.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>