schema-recipe.rng: add element ovs_bridge
This element represents an Open vSwitch bridge. This element has a
mandatory child element <slaves> that defines which previously defined
interfaces should be connected to the new bridge as ports.

To create VLANs on this bridge we can use the <vlan> elements. This vlan
element is different from the element that defines a linux vlan
interface. This element defines a virtual lan identified by it's
"tag" (a mandatory attribute) that contains ports that are defined in
the <slaves> element in the usual way. It is important to note that one
port can be in multiple vlans, this makes the port a VLAN trunk.

Finally the bridge can define zero or more bonds by child elements
<bond>. These are similar to normal linux bonds, however at the moment
they don't accept any options, only the specification of it's slaves.
These bond slaves MUST be a subset of the bridges slaves!

As usual the <define> tag can be used anywhere.

An example use in a recipe:
<ovs_bridge id="ovs_br">
    <slaves>
        <slave id="t1"/>
        <slave id="t2"/>
    </slaves>
    <vlan tag="1">
        <slaves>
            <slave id="t1"/>
        </slaves>
    </vlan>
    <bond id="mfg">
        <slaves>
            <slave id="t1"/>
            <slave id="t2"/>
        </slaves>
    </bond>
</ovs_bridge>

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>