RecipeParser: add support for ovs_bridge element
The slaves of the ovs_bridge are translated to the internal
representation in the same way as slaves of other devices. However
information about vlan tags and bonds is stored in a new "ovs_conf"
field. I added this field to make it relatively simple to add new
features to the recipe schema of ovs_bridges.

In addition to this, the new code checks for conflicts such as
definition of the same bond, enslaving the same port in different bonds
or referencing non-existing ports.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>