{Nm, Net}ConfigDevice: add support for OvS bridges
This commit adds the classes NmConfigDeviceOvsBridge and
NetConfigDeviceOvsBridge that represent the configuration of an Open
vSwitch bridge. The Nm class only defines the is_nm_managed method that
is set to return False and check the devices slaves, since NM does not
support OvS configuration.

The Net class takes care of the OvS configuration by calling appropriate
shell commands, at the moment it can handle the creation and destruction
of an OvS bridge, add/remove ports and set their vlan tags and finally
add/remove bonds. This is sufficient for basic support of Open vSwitch.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>