add ovs_vlan.xml recipe
This commit adds a new example recipe for Open vSwitch. The test follows
the example from the official Open vSwitch Configuration Cookbook on
VLANs
http://openvswitch.org/support/config-cookbooks/vlan-configuration-cookbook/

The only difference is that the recipe uses 6 machines (in my case all
virtual) instead of 2 physical hosts with 2 virtual machines each, but
the end result should be the same.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>