Config: add xslt_url option
This is a Controller specific option, the value is a path or a URL to
the XSLT document that should be referenced in a result xml file if the
option -x is supplied.

The default value points to the xml_to_html.xsl file in this repository
but this should change shortly when we have the files uploaded
somewhere.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>