InterfaceManager: reconstruct upper/lower hierarchy
This patch makes the internal Device objects reconstruct the upper/lower
device hierarchy that can be found in the kernel. This is usefull when
we are deconfiguring devices without information from the kernel - the
upper and lower references will ensure the proper ordering of
deconfiguration.

Internal variables are named like on the controller:
upper == master
lower == slaves
master["primary"] is the primary master of an interface (IFLA_MASTER)

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>