Restructure team recipes
This patches cleans up the team driver recipes directory. Interfaces xml files
were moved to ./interfaces and sequences to ./tasks.

One of the reasons why I make this change was that lnst-ctl is able to run
all recipes under directory if this directory is passed as recipe parameter.
But the mix of recipes, interfaces, sequence was making this impossible.

Additionally I've removed the obsolete machineconfigs that are not used
anywhere anymore.

I've tested this change by running:
./lnst-ctl -d run recipes/team/

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>