TaskAPI: don't use kwargs for passing options to test module
Currently the TaskAPI uses kwargs to pass options to a test module. If a
user decides to run test module from XML there are no limitations on the
name of an option. But when the user uses TaskAPI get_module method the
option names are limited because of Python reserved words [1]. Developer
of a test might want to use e.g. 'if' for a variable containing an interface
to use for testing. To be consistent between the two approaches I'm sending
a patch that uses parameter 'options' of type dictionary for passing the
test module options.

Example:

Old: my_mod = ctl.get_module("Multicast",
                             setup="send_simple",
                             duration=10,
                             condition="status == 'pass'")

New: my_mod = ctl.get_module("Multicast", options={
                                 "setup": "send_simple",
                                 "duration": 10,
                                 "condition": "status == 'pass'"})

I have tested the patch using Multicast test module and it works fine.

[1] https://docs.python.org/2/reference/lexical_analysis.html#keywords

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>