point xslt references to lnst-project.org
This adjusts the generated xml result files to use xslt resources stored
on the lnst-project.org webpage. This should make it very easy to use
xml result files by simply opening them in a web browser which will take
care of formatting them if the computer is connected to the internet.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>