NetTestResultSerializer: add method get_result_html
The new method creates the xml result file that we have been using until
now, reads a specified XSL file (based on the appropriate config option)
and performs a XSLT transformation to create an html file.

The method get_result_xml was renamed to _generate_xml because it is
used by both the old get_result_xml and the new get_result_html methods.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>