NetTestResultSerializer: remove xslt reference in xml
This patch removes the xslt reference that was added to generated xml
files. The original plan was to use this so that the browser can
automatically transform the xml file to a html file. We later discovered
that this is not possible in modern browsers due to security issues of
referencing documents from foreign domains. The reference is therefore
useless and removed.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>