remove redundant interrupt code from PacketAssert test module
Because of using os.killpg(), the process group variant of kill, for
background process termination the interrupt handler in PacketAssert
test module lacks its original purpose.

The os.killpg() method sends signal to whole process group including the
tcpdump command running as forked process therefore the interrupt handler
can be removed as it is redundant to send the signal to the process twice.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>