RecipePath: Add exists and resolve methods
These two new methods, exists() and resolve() are needed for proper handling
of python tasks if a recipe is specified as URL.

The exists() method returns True if a file or URL exists, False
otherwise.

The resolve() method returns the filename containing the data. In case
of ordinary file nothing special needs to be done. In case of URL, the
file is downloaded and saved in temporary file.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>