InterfaceManager: change id to device mapping
This patch replaces the id to device mapping with id to if_index
mapping. This removes the need to update the mapping if a Device object
is removed and recreated for the same device. This can happen when
working with network namespaces - moving a device to a different
namespace deletes it in the current namespace.

Because of this we also need a new attribute - _tmp_mapping that maps id
to device for software devices that were not initialized yet and don't
have an if_index assigned to them.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>