NetConfigDevice: bonding use iproute2 when possible
This patch makes the NetConfigDeviceBond class use the ip command for
device creation/destruction and enslavement instead of sysfs. This is
because bond devices created through sysfs are always in the root
network namespace due to a kernel bug.

However on older systems bonding doesn't support netlink so I added an
exception to use sysfs for kernels older than 3.10. I chose 3.10 because
that's the oldest one I tested that worked - it's the current RHEL7
kernel.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>