recipe: network namespace support
This patch extends the recipe XML files to support network namespaces.
This is a pretty simple change:
* all device elemenets (eth, bond, bridge, ...) now support the "netns"
  attribute which can be an arbitrary string
* command elements config, run, kill, intr and wait now also support the
  "netns" attribute

At the moment there are no checks whether or not the value of the
attribute makes sense. The controller will create the namespaces based
on the device configuration so if a command uses a different namespace
it will fail. The same goes for background commands - the corresponding
signal command needs to be run in the same namespace or it will fail.
This probably can be improved in the future.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>