NetTestSlave: remount sysfs in new net namespace
Creating a new network namespace does not change the contents of the
sysfs filesystem which makes it unusable for network configuration
purposes. This patch fixes that by creating a new mount namespace and
remounting the sysfs filesystem.

This also means that we don't need to use iproute2 to create Bonding
interfaces anymore.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>