NetTestSlave: make return_if_netns method recursive
This patch makes the 'return_if_netns' slave method recursive - the
controller always calls it in the root namespace which will reroute the
method to the specific namespace that the device was moved into. In
addition the InterfaceManager class will not remove Device objects if
they're moved to a namespace (_netns attribute != None) or if they have
a configuration object associated with them. This works around a bug on
older kernels that generate RTM_DELLINK messages for slave devices when
removing the master device (e.g. bridge).

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>