Machine: Interface class veth support
This patch adds support for configuring veth interfaces from the
Interface class and its subclass SoftInterface. The most important
changes are in the configure and deconfigure methods.

When configuring a veth interface we call a new slave method
'create_if_pair' that creates and configures both of the paired
interfaces. The first of the configured interfaces also marks the second
one as configured so it skiped.

Deconfiguration follows a similar pattern  by calling the
'deconfigure_if_pair' method.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>