add vti support
This patch adds support for vti devices. There is no configuration
supported beyond purely creating the device. The recipe format is:

<vti id="xyz" netns="aa">
    <options>
        <option name="local" value="1.1.1.1"/>
        <option name="remote" value="1.1.1.2"/>
        <option name="key" value="1"/>
    </options>
</vti>

Where the key option is mandatory and local and remote options are
optional (though I don't know if such a device is usable). The netns
attribute is optional as usual.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>