signal commands no longer use the netns attribute
This patch improves the usability of signal commands by storing
information about running background commands of the machine. The
controller can then use this information to send the signal command to
the correct network namespace without requiring this information from
the user. Because of this the 'netns' attribute (recipe) is no longer
accepted for signal commands.

RecipeParser also sets the 'netns' attribute of the internal command
structure to the default value 'None'.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>